pyproject.toml
[project]
name = "domos-pmea"
version = "0.1.0"
description = "DOMOS Property Manager Email Assistant Test Assignment"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aioimaplib>=2.0.1",
    "aiosmtplib>=4.0.1",
    "asyncio>=3.4.3",
    "langchain-core>=0.3.61",
    "langchain-redis>=0.1.2",
    "pydantic>=2.11.5",
    "pydantic-settings>=2.9.1",
    "pyyaml>=6.0.2",
    "redis>=6.1.0",
    "uv>=0.7.8",
]

[project.scripts]
domos-pmea = "pmea.main:main"
